Minor fixes
In all, an excellent doc @vsramanujan. It was fun recollecting all this at depth.However, in `generators`, you mention `coroutine` but do not implement it. I tried refactoring that code with out losing the context but couldn't. It would be nice if you provide a working code so that readers can copy +paste to play around.